refactor(settings): merge ClientTagModule into SettingsModule
Other settings bindings (display mode, theme, language, bootstrap relays)
are all provided by SettingsModule. Client tag bindings belong there too.
